chore: move extract-mensa-dishes to repository folder
